Rename to Hentai-Reader and add sidebar
Rename project and artifacts from e-hentai-plus to Hentai-Reader (package.json, package-lock.json, README(s), CI release names, and build output). Update vite config metadata and release URLs. Replace the old single-page scrollbar/thumbnail-panel with a new sidebar component (virtualized thumbnail list + progress track), update overlay code to use the sidebar and wake the progress bar on wheel events, and improve image loading by using a transparent GIF placeholder and CSS fade handling. Remove deprecated scrollbar/thumbnail files and adjust styles accordingly. Version bumped to 2.6.0.
