+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "e-hentai-plus",
+  "name": "hentai-reader",
   "private": true,
-  "version": "2.5.0",
+  "version": "2.6.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
