+++ src/ui/styles.css
@@ -58,27 +58,29 @@ html.scroll-mode #gdt { display: flex; flex-direction: column; align-items: cent
 .interval-input:focus { outline: none; border-color: #888; }
 
 /* Single page overlay */
-.single-page-overlay { position: fixed; top: 0; left: 0; width: 100vw; height: 100vh; background: #000; z-index: 9998; display: none; align-items: center; justify-content: center; }
+.single-page-overlay { position: fixed; top: 0; right: 0; bottom: 0; left: 0; background: #000; z-index: 9998; display: none; align-items: center; justify-content: center; }
 .single-page-overlay.active { display: flex; }
 .sp-image-container { width: 100%; height: 100%; display: flex; align-items: center; justify-content: center; position: relative; }
-.sp-current-image { width: 100%; height: 100%; object-fit: contain; user-select: none; }
+.sp-current-image { width: 100%; height: 100%; object-fit: contain; user-select: none; transition: opacity 0.1s; }
+.sp-current-image:not([src]), .sp-current-image[src=""] { opacity: 0; }
 .sp-close-btn { position: absolute; top: 20px; right: 20px; width: 40px; height: 40px; background: rgba(51,51,51,0.8); border-radius: 50%; display: flex; align-items: center; justify-content: center; cursor: pointer; font-size: 24px; color: #fff; transition: all 0.3s; z-index: 10; }
 .sp-close-btn:hover { background: rgba(85,85,85,0.9); transform: scale(1.1); }
 
-/* Single page scrollbar */
-.sp-scrollbar { position: absolute; right: 40px; top: 10%; width: 12px; height: 80%; background: rgba(40,40,40,0.3); border-radius: 6px; z-index: 10; transition: background 0.3s; }
-.sp-scrollbar:hover { background: rgba(50,50,50,0.5); }
-.sp-scrollbar-thumb { position: absolute; left: 0; width: 100%; min-height: 60px; background: rgba(255,255,255,0.4); border-radius: 6px; transition: background 0.3s; cursor: grab; user-select: none; }
-.sp-scrollbar-thumb:hover { background: rgba(255,255,255,0.6); }
-.sp-scrollbar-thumb:active { cursor: grabbing; background: rgba(255,255,255,0.7); }
-.sp-scrollbar-label { position: absolute; right: calc(100% + 16px); top: 50%; transform: translateY(-50%); background: rgba(26,26,26,0.95); padding: 8px 14px; border-radius: 8px; color: #fff; font-family: monospace; font-size: 14px; white-space: nowrap; opacity: 0; pointer-events: none; transition: opacity 0.3s; box-shadow: 0 2px 8px rgba(0,0,0,0.3); }
-.sp-scrollbar:hover .sp-scrollbar-label { opacity: 1; }
+/* Sidebar Components */
+.sp-sidebar-track { position: absolute; right: 12px; top: 10%; width: 12px; height: 80%; background: rgba(40,40,40,0.3); border-radius: 6px; z-index: 10; pointer-events: none; opacity: 0; transition: opacity 0.3s, background 0.3s; }
+.sp-sidebar-track.active { opacity: 1; pointer-events: auto; }
+.sp-sidebar-track:hover { background: rgba(50,50,50,0.5); }
 
-/* Thumbnail panel */
-.sp-thumb-panel { position: absolute; right: calc(100% + 12px); top: 50%; transform: translateY(-50%); width: 116px; background: rgba(20,20,20,0.95); border-radius: 8px; border: 1px solid rgba(255,255,255,0.1); box-shadow: 0 4px 16px rgba(0,0,0,0.5); opacity: 0; pointer-events: none; transition: opacity 0.2s; z-index: 11; overflow: hidden; }
-.sp-scrollbar:hover .sp-thumb-panel, .sp-thumb-panel:hover { opacity: 1; pointer-events: auto; }
-.sp-scrollbar::before { content: ''; position: absolute; right: 100%; top: 0; width: 140px; height: 100%; }
-.sp-thumb-viewport { width: 100%; max-height: 80vh; overflow: hidden; position: relative; }
+.sp-sidebar-thumb { position: absolute; left: 0; width: 100%; min-height: 60px; background: rgba(255,255,255,0.4); border-radius: 6px; transition: background 0.3s; cursor: grab; user-select: none; }
+.sp-sidebar-thumb:hover { background: rgba(255,255,255,0.6); }
+.sp-sidebar-thumb:active { cursor: grabbing; background: rgba(255,255,255,0.7); }
+
+.sp-sidebar-label { position: absolute; right: calc(100% + 16px); top: 50%; transform: translateY(-50%); background: rgba(26,26,26,0.95); padding: 8px 14px; border-radius: 8px; color: #fff; font-family: monospace; font-size: 14px; white-space: nowrap; opacity: 0; pointer-events: none; transition: opacity 0.3s; box-shadow: 0 2px 8px rgba(0,0,0,0.3); }
+.sp-sidebar-track:hover .sp-sidebar-label { opacity: 1; }
+
+.sp-thumb-panel { position: absolute; right: 12px; top: 10%; max-height: 80%; height: fit-content; width: 116px; background: rgba(20,20,20,0.95); border-radius: 8px; border: 1px solid rgba(255,255,255,0.1); box-shadow: -4px 4px 16px rgba(0,0,0,0.5); overflow: hidden; pointer-events: none; display: flex; flex-direction: column; z-index: 11; opacity: 0; transform: translateX(20px); transition: opacity 0.3s, transform 0.3s cubic-bezier(0.2, 0.8, 0.2, 1); }
+.sp-thumb-panel.active { opacity: 1; transform: translateX(0); pointer-events: auto; }
+.sp-thumb-viewport { flex: 1; min-height: 0; width: 100%; overflow: hidden; position: relative; }
 .sp-thumb-content { position: relative; width: 100%; }
 .sp-thumb-item { position: absolute; left: 8px; width: 100px; height: 72px; border-radius: 4px; overflow: hidden; cursor: pointer; border: 2px solid transparent; transition: border-color 0.15s; box-sizing: border-box; }
 .sp-thumb-item:hover { border-color: rgba(255,255,255,0.4); }
