+++ vite.config.ts
@@ -9,19 +9,19 @@ export default defineConfig({
       entry: 'src/main.ts',
       userscript: {
         name: {
-          '': 'e-hentai Plus',
-          'zh-CN': 'E-Hentai 增强阅读',
+          '': 'Hentai Reader',
+          'zh-CN': '绅士阅读器',
         },
         namespace: 'http://tampermonkey.net/',
-        homepageURL: 'https://github.com/Leovikii/e-hentai-plus',
+        homepageURL: 'https://github.com/Leovikii/Hentai-Reader',
 
         description: {
           '': 'Infinite scroll & reader mode with image prefetch and floating controls for E-Hentai / ExHentai',
           'zh-CN': '为 E-Hentai / ExHentai 提供无限滚动和阅读模式，支持图片预取与悬浮控制',
         },
         author: 'Viki',
-        updateURL: 'https://github.com/Leovikii/e-hentai-plus/releases/latest/download/e-hentai-plus.user.js',
-        downloadURL: 'https://github.com/Leovikii/e-hentai-plus/releases/latest/download/e-hentai-plus.user.js',
+        updateURL: 'https://github.com/Leovikii/Hentai-Reader/releases/latest/download/hentai-reader.user.js',
+        downloadURL: 'https://github.com/Leovikii/Hentai-Reader/releases/latest/download/hentai-reader.user.js',
         match: [
           'https://e-hentai.org/g/*',
           'https://exhentai.org/g/*',
@@ -34,7 +34,7 @@ export default defineConfig({
         license: 'MIT',
       },
       build: {
-        fileName: 'e-hentai-plus.user.js',
+        fileName: 'hentai-reader.user.js',
         autoGrant: true,
       },
     }),
